requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
certifi==2023.11.17
cfgv==3.4.0
charset-normalizer==3.3.2
distlib==0.3.8
Django==4.2
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.24.1
Faker==26.0.0
filelock==3.13.1
gunicorn==21.2.0
identify==2.5.33
idna==3.6
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
nodeenv==1.8.0
packaging==23.2
Pillow==10.1.0
platformdirs==4.1.0
pre-commit==3.5.0
psycopg2-binary==2.9.9
py-code-meli==0.1.3
PyJWT==2.8.0
pymemcache==4.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.32.0
requests==2.31.0
rpds-py==0.13.2
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
virtualenv==20.25.0
